+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file. The format 
 - Avatar, Box, Button, and Checkbox pages now use the drawer
 - Modal, Pagination, Panel, Progress, TextForm, Radio Group, Slider,
   Select, Snackbar, Switch, Tooltip, and Typography pages now use the drawer
+- AppBar, SpeedDial, Stepper, Tree, and Video pages now use the drawer
 
 ## [v0.8.5]
 ### Added